fix subtle window function bug, `LastNotNullValueOverRowsFrameFunction` would initialise lastValue=0 (default), but then reset it to lastValue=null. also, cacheValue was not being reset between runs
