+++ core/src/test/java/io/questdb/test/griffin/engine/groupby/SampleByTest.java
@@ -157,6 +157,358 @@ public class SampleByTest extends AbstractCairoTest {
     }
 
     @Test
+    public void testFillPrevConsistency() throws SqlException {
+        execute("create table telem (created timestamp, event_type int, table_id int, latency double) timestamp(created) partition by DAY");
+
+        execute("insert into telem " +
+                "select" +
+                " timestamp_sequence('2025-01-20T13:57:14.000000Z'::timestamp, 2500)," +
+                " rnd_int() % 4," +
+                " rnd_int() % 100," +
+                " abs(rnd_double())" +
+                " from long_sequence(100_000)");
+
+        // todo: return values for "prev" function are incorrect. This is a bug.
+        // [nwoolmer] I think I've fixed this now, dirty state in SimpleMapValuePeeker
+        assertQueryNoLeakCheck(
+                "created\tavg\tlatency\n" +
+                        "2025-01-20T13:56:50.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:56:52.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:56:54.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:56:56.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:56:58.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:00.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:02.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:04.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:06.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:08.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:10.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:12.000000Z\t0.0\t0.0\n" +
+                        "2025-01-20T13:57:14.000000Z\t0.4851638802935891\t0.4846019644078461\n" +
+                        "2025-01-20T13:57:16.000000Z\t0.5040684715238979\t0.0014510055926236776\n" +
+                        "2025-01-20T13:57:18.000000Z\t0.4855058436740148\t0.760595244599882\n" +
+                        "2025-01-20T13:57:20.000000Z\t0.5022292093071415\t0.47033911949468266\n" +
+                        "2025-01-20T13:57:22.000000Z\t0.5052664001761066\t0.8733949114345736\n" +
+                        "2025-01-20T13:57:24.000000Z\t0.49787296387399244\t0.45485294118316344\n" +
+                        "2025-01-20T13:57:26.000000Z\t0.5157930891201031\t0.7007778377722195\n" +
+                        "2025-01-20T13:57:28.000000Z\t0.49810868200578157\t0.22896810382600008\n" +
+                        "2025-01-20T13:57:30.000000Z\t0.4939448789648544\t0.77359848960926\n" +
+                        "2025-01-20T13:57:32.000000Z\t0.4897555120505765\t0.8994724612954251\n" +
+                        "2025-01-20T13:57:34.000000Z\t0.5017879450710985\t0.5496601327901486\n" +
+                        "2025-01-20T13:57:36.000000Z\t0.4956674689326836\t0.26110257790188807\n" +
+                        "2025-01-20T13:57:38.000000Z\t0.5108794030479952\t0.38345997979929847\n" +
+                        "2025-01-20T13:57:40.000000Z\t0.5053240993230795\t0.520757673455773\n" +
+                        "2025-01-20T13:57:42.000000Z\t0.4988663699172193\t0.7881958875290708\n" +
+                        "2025-01-20T13:57:44.000000Z\t0.503737560421611\t0.4669565407466736\n" +
+                        "2025-01-20T13:57:46.000000Z\t0.5099543384735293\t0.5608039955756913\n" +
+                        "2025-01-20T13:57:48.000000Z\t0.4959893946518482\t0.2843815949074908\n" +
+                        "2025-01-20T13:57:50.000000Z\t0.5016936986983442\t0.8505737925875698\n" +
+                        "2025-01-20T13:57:52.000000Z\t0.5054365351895309\t0.12286408407144689\n" +
+                        "2025-01-20T13:57:54.000000Z\t0.4906196065427902\t0.18333103699732645\n" +
+                        "2025-01-20T13:57:56.000000Z\t0.5000897294717391\t0.8657117894806349\n" +
+                        "2025-01-20T13:57:58.000000Z\t0.5098086517700594\t0.34100284129451275\n" +
+                        "2025-01-20T13:58:00.000000Z\t0.49966159558186946\t0.12147807024033108\n" +
+                        "2025-01-20T13:58:02.000000Z\t0.5004546880326035\t0.6448153122398131\n" +
+                        "2025-01-20T13:58:04.000000Z\t0.509048504935009\t0.36592352333201006\n" +
+                        "2025-01-20T13:58:06.000000Z\t0.50368892165536\t0.8537660947780769\n" +
+                        "2025-01-20T13:58:08.000000Z\t0.5135177208750468\t0.9820930620204293\n" +
+                        "2025-01-20T13:58:10.000000Z\t0.4946776746792163\t0.5371999576078124\n" +
+                        "2025-01-20T13:58:12.000000Z\t0.5046786160937665\t0.4597099533579079\n" +
+                        "2025-01-20T13:58:14.000000Z\t0.5104309717281234\t0.9703717493026135\n" +
+                        "2025-01-20T13:58:16.000000Z\t0.5063263221664347\t0.3173787005052606\n" +
+                        "2025-01-20T13:58:18.000000Z\t0.5050422491444898\t0.8321198770692697\n" +
+                        "2025-01-20T13:58:20.000000Z\t0.49942269756582924\t0.15080308349106386\n" +
+                        "2025-01-20T13:58:22.000000Z\t0.5001306567656737\t0.0754281163419458\n" +
+                        "2025-01-20T13:58:24.000000Z\t0.49295208178613864\t0.9199510142446139\n" +
+                        "2025-01-20T13:58:26.000000Z\t0.5059398643771146\t0.4660874688287775\n" +
+                        "2025-01-20T13:58:28.000000Z\t0.4981379556559048\t0.30033880795051826\n" +
+                        "2025-01-20T13:58:30.000000Z\t0.5000418143092644\t0.5705182422010522\n" +
+                        "2025-01-20T13:58:32.000000Z\t0.5106167257086114\t0.09515699608123174\n" +
+                        "2025-01-20T13:58:34.000000Z\t0.5067174084443764\t0.05834219734998902\n" +
+                        "2025-01-20T13:58:36.000000Z\t0.5068112645934018\t0.1817699293870746\n" +
+                        "2025-01-20T13:58:38.000000Z\t0.4988092223173822\t0.7193513157733163\n" +
+                        "2025-01-20T13:58:40.000000Z\t0.5022197872758152\t0.718094867327548\n" +
+                        "2025-01-20T13:58:42.000000Z\t0.5106127706294329\t0.8837745591581858\n" +
+                        "2025-01-20T13:58:44.000000Z\t0.5028917099020589\t0.47476695890232523\n" +
+                        "2025-01-20T13:58:46.000000Z\t0.5007363609737706\t0.13736027364505599\n" +
+                        "2025-01-20T13:58:48.000000Z\t0.492334610467138\t0.5229151773456188\n" +
+                        "2025-01-20T13:58:50.000000Z\t0.5001068629548213\t0.5240982714764488\n" +
+                        "2025-01-20T13:58:52.000000Z\t0.4994174463473467\t0.47333881895683316\n" +
+                        "2025-01-20T13:58:54.000000Z\t0.49424015140261957\t0.42701619170227834\n" +
+                        "2025-01-20T13:58:56.000000Z\t0.48757211780610227\t0.08512812943026549\n" +
+                        "2025-01-20T13:58:58.000000Z\t0.5100223179690243\t0.16819110732876985\n" +
+                        "2025-01-20T13:59:00.000000Z\t0.49943191308339374\t0.2702475064009233\n" +
+                        "2025-01-20T13:59:02.000000Z\t0.48652969691916503\t0.7998124033479126\n" +
+                        "2025-01-20T13:59:04.000000Z\t0.5016355069527151\t0.7227406508693206\n" +
+                        "2025-01-20T13:59:06.000000Z\t0.5072743308494192\t0.8646291312820938\n" +
+                        "2025-01-20T13:59:08.000000Z\t0.5146616087916882\t0.593463629527755\n" +
+                        "2025-01-20T13:59:10.000000Z\t0.4924321489143956\t0.9113233365891048\n" +
+                        "2025-01-20T13:59:12.000000Z\t0.5019611647508408\t0.5287400406602393\n" +
+                        "2025-01-20T13:59:14.000000Z\t0.5047832584945308\t0.7358918625777293\n" +
+                        "2025-01-20T13:59:16.000000Z\t0.506063752855275\t0.5402660817376952\n" +
+                        "2025-01-20T13:59:18.000000Z\t0.524139855411055\t0.6610279878811472\n" +
+                        "2025-01-20T13:59:20.000000Z\t0.49536698884043756\t0.8383151117609724\n" +
+                        "2025-01-20T13:59:22.000000Z\t0.523861275985825\t0.4449315194100708\n" +
+                        "2025-01-20T13:59:24.000000Z\t0.5065803411918478\t0.33950559953239867\n" +
+                        "2025-01-20T13:59:26.000000Z\t0.49505000311232045\t0.36271880845750437\n" +
+                        "2025-01-20T13:59:28.000000Z\t0.49762813072208195\t0.7118659078648814\n" +
+                        "2025-01-20T13:59:30.000000Z\t0.4914193450153209\t0.22370121110446894\n" +
+                        "2025-01-20T13:59:32.000000Z\t0.49997792522902584\t0.12484119317638909\n" +
+                        "2025-01-20T13:59:34.000000Z\t0.5131846146411408\t0.7627858499872365\n" +
+                        "2025-01-20T13:59:36.000000Z\t0.50575819056176\t0.2989367055630815\n" +
+                        "2025-01-20T13:59:38.000000Z\t0.5115672301835786\t0.7311461454026085\n" +
+                        "2025-01-20T13:59:40.000000Z\t0.489770739523703\t0.364203755015714\n" +
+                        "2025-01-20T13:59:42.000000Z\t0.48827098190254603\t0.6497666544419626\n" +
+                        "2025-01-20T13:59:44.000000Z\t0.4919096456204459\t0.1751598713431003\n" +
+                        "2025-01-20T13:59:46.000000Z\t0.5025937922857477\t0.7887957915957375\n" +
+                        "2025-01-20T13:59:48.000000Z\t0.4802568230380409\t0.7425480329713459\n" +
+                        "2025-01-20T13:59:50.000000Z\t0.5052745501212974\t0.7105531082982322\n" +
+                        "2025-01-20T13:59:52.000000Z\t0.5153544761406945\t0.356986250439874\n" +
+                        "2025-01-20T13:59:54.000000Z\t0.5226395082273337\t0.6793175726750598\n" +
+                        "2025-01-20T13:59:56.000000Z\t0.5113708824401949\t0.6917192181729446\n" +
+                        "2025-01-20T13:59:58.000000Z\t0.5007369251682845\t0.5566259650060899\n" +
+                        "2025-01-20T14:00:00.000000Z\t0.4875623932277653\t0.41544865478057724\n" +
+                        "2025-01-20T14:00:02.000000Z\t0.500581587909613\t0.35535692290002574\n" +
+                        "2025-01-20T14:00:04.000000Z\t0.5033365462753482\t0.6425242900479169\n" +
+                        "2025-01-20T14:00:06.000000Z\t0.49973458951836724\t0.8057830760536094\n" +
+                        "2025-01-20T14:00:08.000000Z\t0.5133768100928325\t0.4832635039849207\n" +
+                        "2025-01-20T14:00:10.000000Z\t0.49959821200212373\t0.07969074155053768\n" +
+                        "2025-01-20T14:00:12.000000Z\t0.48631574988877824\t0.1852594496810288\n" +
+                        "2025-01-20T14:00:14.000000Z\t0.4775700508803147\t0.1746040305540778\n" +
+                        "2025-01-20T14:00:16.000000Z\t0.4904731973923076\t0.8319077385303282\n" +
+                        "2025-01-20T14:00:18.000000Z\t0.5172686285942334\t0.4602012941029604\n" +
+                        "2025-01-20T14:00:20.000000Z\t0.49438245747522913\t0.496679499909753\n" +
+                        "2025-01-20T14:00:22.000000Z\t0.48890769363602105\t0.6624975451169404\n" +
+                        "2025-01-20T14:00:24.000000Z\t0.524539276436426\t0.7613314616414868\n" +
+                        "2025-01-20T14:00:26.000000Z\t0.47754365417732036\t0.7972715048962937\n" +
+                        "2025-01-20T14:00:28.000000Z\t0.49558490815429723\t0.5650524947355062\n" +
+                        "2025-01-20T14:00:30.000000Z\t0.478522476906104\t0.8980004790312893\n" +
+                        "2025-01-20T14:00:32.000000Z\t0.5032895135257704\t0.8725153533702712\n" +
+                        "2025-01-20T14:00:34.000000Z\t0.4885618841436868\t0.16756071058522382\n" +
+                        "2025-01-20T14:00:36.000000Z\t0.4945189248409826\t0.20602018544449763\n" +
+                        "2025-01-20T14:00:38.000000Z\t0.49157255178162024\t0.9762360584775526\n" +
+                        "2025-01-20T14:00:40.000000Z\t0.5021366595107085\t2.2543219813986326E-4\n" +
+                        "2025-01-20T14:00:42.000000Z\t0.5010926084080161\t0.9927756543484265\n" +
+                        "2025-01-20T14:00:44.000000Z\t0.5061735021344332\t0.3034233576280241\n" +
+                        "2025-01-20T14:00:46.000000Z\t0.49349786391999323\t0.2990105656288151\n" +
+                        "2025-01-20T14:00:48.000000Z\t0.5052800074403885\t0.1118642844938007\n" +
+                        "2025-01-20T14:00:50.000000Z\t0.478761125271968\t0.6763386750115261\n" +
+                        "2025-01-20T14:00:52.000000Z\t0.514706027105563\t0.7763367804571444\n" +
+                        "2025-01-20T14:00:54.000000Z\t0.5055529298956492\t0.29442457234943387\n" +
+                        "2025-01-20T14:00:56.000000Z\t0.4750145366710726\t0.20393005251349205\n" +
+                        "2025-01-20T14:00:58.000000Z\t0.4884816249911365\t0.33499512944932763\n" +
+                        "2025-01-20T14:01:00.000000Z\t0.5094157389039983\t0.021272791876005726\n" +
+                        "2025-01-20T14:01:02.000000Z\t0.5047514008290243\t0.5788643390071768\n" +
+                        "2025-01-20T14:01:04.000000Z\t0.4863102272143384\t0.00965027758091308\n" +
+                        "2025-01-20T14:01:06.000000Z\t0.5071935666235293\t0.762329563481065\n" +
+                        "2025-01-20T14:01:08.000000Z\t0.5088687556412043\t0.28544640285844014\n" +
+                        "2025-01-20T14:01:10.000000Z\t0.4960479410464751\t0.49922336302934456\n" +
+                        "2025-01-20T14:01:12.000000Z\t0.4979687963161959\t0.8774491669657007\n" +
+                        "2025-01-20T14:01:14.000000Z\t0.4959509451344056\t0.37826342574379146\n" +
+                        "2025-01-20T14:01:16.000000Z\t0.51067644621747\t0.7751692205906762\n" +
+                        "2025-01-20T14:01:18.000000Z\t0.49924129070780227\t0.5021223982023798\n" +
+                        "2025-01-20T14:01:20.000000Z\t0.4972046530980123\t0.5263994817020877\n" +
+                        "2025-01-20T14:01:22.000000Z\t0.5080393379955584\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:24.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:26.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:28.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:30.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:32.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:34.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:36.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:38.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:40.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:42.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:44.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:46.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:48.000000Z\t0.0\t0.26648957155568953\n" +
+                        "2025-01-20T14:01:50.000000Z\t0.0\t0.26648957155568953\n",
+                "select created, avg(latency) avg, last(latency) latency\n" +
+                        "  from telem\n" +
+                        "  sample by 2s\n" +
+                        "  FROM timestamp_floor('2s', '2025-01-20T13:56:50Z')\n" +
+                        "  TO timestamp_floor('2s', '2025-01-20T14:01:52Z')\n" +
+                        "  fill(0,prev);",
+                "created"
+        );
+    }
+
+    @Test
+    public void testFillValueException() throws SqlException {
+        execute("create table telem (created timestamp, event_type int, table_id int, latency double) timestamp(created) partition by DAY");
+
+        execute("insert into telem " +
+                "select " +
+                "generate_series as created," +
+                " rnd_int() % 4," +
+                " rnd_int() % 100," +
+                " abs(rnd_double())" +
+                " from generate_series('2025-01-20T13:57:14.000000Z', dateadd('u', 2500 * 100_000, '2025-01-20T13:57:14.000000Z') - 1, 2500)");
+
+        assertQueryNoLeakCheck(
+                "created\tlatency\n" +
+                        "2025-01-20T13:56:50.000000Z\t0.0\n" +
+                        "2025-01-20T13:56:52.000000Z\t0.0\n" +
+                        "2025-01-20T13:56:54.000000Z\t0.0\n" +
+                        "2025-01-20T13:56:56.000000Z\t0.0\n" +
+                        "2025-01-20T13:56:58.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:00.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:02.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:04.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:06.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:08.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:10.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:12.000000Z\t0.0\n" +
+                        "2025-01-20T13:57:14.000000Z\t0.8745107650756836\n" +
+                        "2025-01-20T13:57:16.000000Z\t0.8901357650756836\n" +
+                        "2025-01-20T13:57:18.000000Z\t0.8872051239013672\n" +
+                        "2025-01-20T13:57:20.000000Z\t0.8901366591453552\n" +
+                        "2025-01-20T13:57:22.000000Z\t0.9111318588256836\n" +
+                        "2025-01-20T13:57:24.000000Z\t0.9057612419128418\n" +
+                        "2025-01-20T13:57:26.000000Z\t0.9077143669128418\n" +
+                        "2025-01-20T13:57:28.000000Z\t0.8959951400756836\n" +
+                        "2025-01-20T13:57:30.000000Z\t0.9023432731628418\n" +
+                        "2025-01-20T13:57:32.000000Z\t0.8984373807907104\n" +
+                        "2025-01-20T13:57:34.000000Z\t0.9057607650756836\n" +
+                        "2025-01-20T13:57:36.000000Z\t0.8852519989013672\n" +
+                        "2025-01-20T13:57:38.000000Z\t0.9028315544128418\n" +
+                        "2025-01-20T13:57:40.000000Z\t0.8979473114013672\n" +
+                        "2025-01-20T13:57:42.000000Z\t0.9116191864013672\n" +
+                        "2025-01-20T13:57:44.000000Z\t0.9096670150756836\n" +
+                        "2025-01-20T13:57:46.000000Z\t0.9145498275756836\n" +
+                        "2025-01-20T13:57:48.000000Z\t0.8930625915527344\n" +
+                        "2025-01-20T13:57:50.000000Z\t0.8911123275756836\n" +
+                        "2025-01-20T13:57:52.000000Z\t0.8969717025756836\n" +
+                        "2025-01-20T13:57:54.000000Z\t0.8632811903953552\n" +
+                        "2025-01-20T13:57:56.000000Z\t0.9052729606628418\n" +
+                        "2025-01-20T13:57:58.000000Z\t0.9042966365814209\n" +
+                        "2025-01-20T13:58:00.000000Z\t0.8984370231628418\n" +
+                        "2025-01-20T13:58:02.000000Z\t0.9038066864013672\n" +
+                        "2025-01-20T13:58:04.000000Z\t0.8984372615814209\n" +
+                        "2025-01-20T13:58:06.000000Z\t0.8872067928314209\n" +
+                        "2025-01-20T13:58:08.000000Z\t0.9145488739013672\n" +
+                        "2025-01-20T13:58:10.000000Z\t0.8940410614013672\n" +
+                        "2025-01-20T13:58:12.000000Z\t0.9042949676513672\n" +
+                        "2025-01-20T13:58:14.000000Z\t0.9023427963256836\n" +
+                        "2025-01-20T13:58:16.000000Z\t0.9086909294128418\n" +
+                        "2025-01-20T13:58:18.000000Z\t0.9023432731628418\n" +
+                        "2025-01-20T13:58:20.000000Z\t0.9062480926513672\n" +
+                        "2025-01-20T13:58:22.000000Z\t0.9082030057907104\n" +
+                        "2025-01-20T13:58:24.000000Z\t0.8999004364013672\n" +
+                        "2025-01-20T13:58:26.000000Z\t0.8969721794128418\n" +
+                        "2025-01-20T13:58:28.000000Z\t0.8984355926513672\n" +
+                        "2025-01-20T13:58:30.000000Z\t0.9135732650756836\n" +
+                        "2025-01-20T13:58:32.000000Z\t0.9082026481628418\n" +
+                        "2025-01-20T13:58:34.000000Z\t0.8989248275756836\n" +
+                        "2025-01-20T13:58:36.000000Z\t0.8916015475988388\n" +
+                        "2025-01-20T13:58:38.000000Z\t0.8940420150756836\n" +
+                        "2025-01-20T13:58:40.000000Z\t0.8886718747671694\n" +
+                        "2025-01-20T13:58:42.000000Z\t0.8945307731628418\n" +
+                        "2025-01-20T13:58:44.000000Z\t0.9101561903953552\n" +
+                        "2025-01-20T13:58:46.000000Z\t0.8999013900756836\n" +
+                        "2025-01-20T13:58:48.000000Z\t0.8935545682907104\n" +
+                        "2025-01-20T13:58:50.000000Z\t0.9042967557907104\n" +
+                        "2025-01-20T13:58:52.000000Z\t0.9033198356628418\n" +
+                        "2025-01-20T13:58:54.000000Z\t0.9067382216453552\n" +
+                        "2025-01-20T13:58:56.000000Z\t0.8867177963256836\n" +
+                        "2025-01-20T13:58:58.000000Z\t0.9077129364013672\n" +
+                        "2025-01-20T13:59:00.000000Z\t0.9052729606628418\n" +
+                        "2025-01-20T13:59:02.000000Z\t0.8964824676513672\n" +
+                        "2025-01-20T13:59:04.000000Z\t0.8920896053314209\n" +
+                        "2025-01-20T13:59:06.000000Z\t0.9077138900756836\n" +
+                        "2025-01-20T13:59:08.000000Z\t0.9096674919128418\n" +
+                        "2025-01-20T13:59:10.000000Z\t0.8901362419128418\n" +
+                        "2025-01-20T13:59:12.000000Z\t0.9067380428314209\n" +
+                        "2025-01-20T13:59:14.000000Z\t0.9086894989013672\n" +
+                        "2025-01-20T13:59:16.000000Z\t0.8930644989013672\n" +
+                        "2025-01-20T13:59:18.000000Z\t0.9077110290527344\n" +
+                        "2025-01-20T13:59:20.000000Z\t0.8979487419128418\n" +
+                        "2025-01-20T13:59:22.000000Z\t0.9350583553314209\n" +
+                        "2025-01-20T13:59:24.000000Z\t0.9091792106628418\n" +
+                        "2025-01-20T13:59:26.000000Z\t0.8920893669128418\n" +
+                        "2025-01-20T13:59:28.000000Z\t0.8867182731628418\n" +
+                        "2025-01-20T13:59:30.000000Z\t0.8950190544128418\n" +
+                        "2025-01-20T13:59:32.000000Z\t0.8994121551513672\n" +
+                        "2025-01-20T13:59:34.000000Z\t0.9130858182907104\n" +
+                        "2025-01-20T13:59:36.000000Z\t0.9013652801513672\n" +
+                        "2025-01-20T13:59:38.000000Z\t0.9033198356628418\n" +
+                        "2025-01-20T13:59:40.000000Z\t0.8998985290527344\n" +
+                        "2025-01-20T13:59:42.000000Z\t0.8920897841453552\n" +
+                        "2025-01-20T13:59:44.000000Z\t0.8989255428314209\n" +
+                        "2025-01-20T13:59:46.000000Z\t0.8959956169128418\n" +
+                        "2025-01-20T13:59:48.000000Z\t0.8784179091453552\n" +
+                        "2025-01-20T13:59:50.000000Z\t0.8984355926513672\n" +
+                        "2025-01-20T13:59:52.000000Z\t0.9086904525756836\n" +
+                        "2025-01-20T13:59:54.000000Z\t0.9018516540527344\n" +
+                        "2025-01-20T13:59:56.000000Z\t0.9023418426513672\n" +
+                        "2025-01-20T13:59:58.000000Z\t0.8989248275756836\n" +
+                        "2025-01-20T14:00:00.000000Z\t0.8847651481628418\n" +
+                        "2025-01-20T14:00:02.000000Z\t0.8945293426513672\n" +
+                        "2025-01-20T14:00:04.000000Z\t0.9111326932907104\n" +
+                        "2025-01-20T14:00:06.000000Z\t0.9052729606628418\n" +
+                        "2025-01-20T14:00:08.000000Z\t0.9160155057907104\n" +
+                        "2025-01-20T14:00:10.000000Z\t0.8911113739013672\n" +
+                        "2025-01-20T14:00:12.000000Z\t0.8881834745407104\n" +
+                        "2025-01-20T14:00:14.000000Z\t0.8852519989013672\n" +
+                        "2025-01-20T14:00:16.000000Z\t0.9125975966453552\n" +
+                        "2025-01-20T14:00:18.000000Z\t0.8911123275756836\n" +
+                        "2025-01-20T14:00:20.000000Z\t0.8911113739013672\n" +
+                        "2025-01-20T14:00:22.000000Z\t0.8754844665527344\n" +
+                        "2025-01-20T14:00:24.000000Z\t0.9018553495407104\n" +
+                        "2025-01-20T14:00:26.000000Z\t0.8959941864013672\n" +
+                        "2025-01-20T14:00:28.000000Z\t0.8920879364013672\n" +
+                        "2025-01-20T14:00:30.000000Z\t0.8730449676513672\n" +
+                        "2025-01-20T14:00:32.000000Z\t0.9077148139476776\n" +
+                        "2025-01-20T14:00:34.000000Z\t0.8945311307907104\n" +
+                        "2025-01-20T14:00:36.000000Z\t0.9057614803314209\n" +
+                        "2025-01-20T14:00:38.000000Z\t0.9125971794128418\n" +
+                        "2025-01-20T14:00:40.000000Z\t0.9091795682907104\n" +
+                        "2025-01-20T14:00:42.000000Z\t0.8955075740814209\n" +
+                        "2025-01-20T14:00:44.000000Z\t0.8847655057907104\n" +
+                        "2025-01-20T14:00:46.000000Z\t0.9169912338256836\n" +
+                        "2025-01-20T14:00:48.000000Z\t0.8974606990814209\n" +
+                        "2025-01-20T14:00:50.000000Z\t0.9023432731628418\n" +
+                        "2025-01-20T14:00:52.000000Z\t0.9160155057907104\n" +
+                        "2025-01-20T14:00:54.000000Z\t0.9038081169128418\n" +
+                        "2025-01-20T14:00:56.000000Z\t0.8588864803314209\n" +
+                        "2025-01-20T14:00:58.000000Z\t0.8999018669128418\n" +
+                        "2025-01-20T14:01:00.000000Z\t0.9042959213256836\n" +
+                        "2025-01-20T14:01:02.000000Z\t0.8916015475988388\n" +
+                        "2025-01-20T14:01:04.000000Z\t0.8964838981628418\n" +
+                        "2025-01-20T14:01:06.000000Z\t0.9028317928314209\n" +
+                        "2025-01-20T14:01:08.000000Z\t0.8867182731628418\n" +
+                        "2025-01-20T14:01:10.000000Z\t0.8945307731628418\n" +
+                        "2025-01-20T14:01:12.000000Z\t0.8979489803314209\n" +
+                        "2025-01-20T14:01:14.000000Z\t0.8964842557907104\n" +
+                        "2025-01-20T14:01:16.000000Z\t0.8916010856628418\n" +
+                        "2025-01-20T14:01:18.000000Z\t0.9023418426513672\n" +
+                        "2025-01-20T14:01:20.000000Z\t0.9038083553314209\n" +
+                        "2025-01-20T14:01:22.000000Z\t0.8930654525756836\n" +
+                        "2025-01-20T14:01:24.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:26.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:28.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:30.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:32.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:34.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:36.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:38.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:40.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:42.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:44.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:46.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:48.000000Z\t0.0\n" +
+                        "2025-01-20T14:01:50.000000Z\t0.0\n",
+                "select created, approx_percentile(latency, 0.9, 3) latency\n" +
+                        "  from telem\n" +
+                        "  sample by 2s\n" +
+                        "  FROM timestamp_floor('2s', '2025-01-20T13:56:50Z')\n" +
+                        "  TO timestamp_floor('2s', '2025-01-20T14:01:52Z')\n" +
+                        "  fill(0);",
+                "created"
+        );
+    }
+
+    @Test
     public void testGeohashFillNull() throws Exception {
         assertQuery(
                 "s\tk\tfirst\tfirst1\tfirst2\tfirst3\n" +
